// src/common/interfaces/nutrition.ts
import { INutrient } from './nutrient';

export interface INutrition {
  PROCNT: INutrient;
  FAT: INutrient;
  CHOCDF: INutrient;
  ENERC_KCAL: INutrient;
  SUGAR: INutrient;
  FIBTG: INutrient;
  CA: INutrient;
  FE: INutrient;
  P: INutrient;
  K: INutrient;
  NA: INutrient;
  VITA_IU: INutrient;
  TOCPHA: INutrient;
  VITD: INutrient;
  VITC: INutrient;
  VITB12: INutrient;
  FOLAC: INutrient;
  CHOLE: INutrient;
  FATRN: INutrient;
  FASAT: INutrient;
  FAMS: INutrient;
  FAPU: INutrient;
}